refactor: move BOARD_CONFIG from types to common
types/ 레이어는 순수 타입만 두는 규칙이라, 런타임 설정(BOARD_CONFIG)과
라벨/설명/권한을 common/board-config.ts로 분리한다. board.type.ts에는
BoardId 타입만 남기고, BOARD_CONFIG import 6곳을 새 경로로 재지정.
동작 변화 없음.
