+++ src/lib/rules/post.rule.ts
@@ -1,7 +1,7 @@
 import type { User } from '$lib/types/user.type.js';
 import type { Post, PostCreate } from '$lib/types/post.type.js';
 import { UserGroup } from '$lib/types/user.type.js';
-import { BOARD_CONFIG } from '$lib/types/board.type.js';
+import { BOARD_CONFIG } from '$lib/common/board-config.js';
 import { hasMinRole } from '$lib/common/permission.js';
 
 export function canEditOrDeletePost(post: Post, user: User): boolean {
