+++ src/routes/board/[boardId=board]/_components/BoardArticle.svelte
@@ -4,7 +4,7 @@
 
 	import type { Post } from '$lib/types/post.type.js';
 	import type { User } from '$lib/types/user.type.js';
-	import { BOARD_CONFIG } from '$lib/types/board.type.js';
+	import { BOARD_CONFIG } from '$lib/common/board-config.js';
 	import type { BoardId } from '$lib/types/board.type.js';
 
 	import * as CommonUtils from '$lib/common/utils.js';
